Extract PlayableQuickAction(Divider) into its own widget file
Moves the per-button widget and the gradient hairline divider out of
playable_action_sheet.dart and into lib/ui/widgets/, exported from
widgets.dart so other sheets can reuse the same quick-action row
layout. The action sheet now references the public types via the
widgets barrel; _CloudMinusIcon stays inline since it's specific to
the download action.
